Preserve project settings on uninstall by default

+++ uninstall.php
@@ -9,6 +9,9 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
+$xpressui_delete_project_settings = defined( 'XPRESSUI_BRIDGE_DELETE_PROJECT_SETTINGS_ON_UNINSTALL' )
+	&& XPRESSUI_BRIDGE_DELETE_PROJECT_SETTINGS_ON_UNINSTALL;
+
 $xpressui_submission_ids = get_posts(
 	[
 		'post_type'      => 'xpressui_submission',
@@ -22,7 +25,9 @@ foreach ( $xpressui_submission_ids as $xpressui_submission_id ) {
 	wp_delete_post( $xpressui_submission_id, true );
 }
 
-delete_option( 'xpressui_project_settings' );
+if ( $xpressui_delete_project_settings ) {
+	delete_option( 'xpressui_project_settings' );
+}
 delete_option( 'xpressui_workflow_manifest_registry' );
 delete_option( 'xpressui_bundled_workflows_installed' );
 delete_option( 'xpressui_bundled_workflows_version' );
